Create 人CRYPTO人
A .scpt file for running in AppleScript that calls on Safari to navigate to an online market price aggregator to pull real-time prices of cryptocurrencies using JavaScript. The script then opens Excel and writes the data to a master workbook for collecting historical information on the stock prices and is used to create trend lines and forecasts with the given data. On my local machine I then used Automator to schedule an alarm that calls the script without user input to run every 15 minutes and then the script closes out Safari and Excel. I used Automator instead of another program because I am learning coding as I go and couldn't figure out how to get the idle handle to run in an Application version of the file. It seems there are too many functions in the 1 file for the idle handle to run uninterrupted in an always running Application version. Currently there is a bug with the script that won't return the month value correctly in Excel that is bothering me and I haven't figured out how to remedy that yet. Working the bugs out as I find them and continuing on the Excel master file with aggregating what I can of the 8,697 cryptocurrencies there are available to buy stock in on https://coinmarketcap.com. I'm not sure if I am going to share the Excel data I collect yet as I spent a fair amount of time and money learning the techniques for assembling the data and generating mathematical models of the price trends. If anyone is interested in the project and the historical and current data I am collecting I would be interested in sharing it for compensation for the time it takes to compile it. The AppleScript coding was learned in my spare time using free resources but the techniques used for forecasting were learned at university, Microsoft Excel is a yearly investment, and my shiny Apple MacBook cost me more than a fancy date in Nevada. I would like to receive some return on that investment before just uploading it all. After all, the information was free to collect and if one is really that interested in the data one could invest time and energy to find their own crystal ball to the crypto-market. Plus I wouldn't want to insinuate that my forecasts are all that amazing to behold. My current progress shows that with the volatility of Bitcoin at least that the models are fairly weak over the long term. I could focus on a trending section of the data but that takes all the confidence level out of forecasting in the first place. Additional work needs to be done collecting market prices and socio-economic data to generate cross-correlation between indicators to find a trends that holds strong and reliable over time to put any confidence level in what appears to be a pretty random free-for-all market bonanza with Bitcoin. I think the best I can hope to do with the script is keep a watchful eye over time, build a familiarity with the data by modeling when I'm free to do it, and possibly taking advantage of a market spike in Bitcoin that would allow me to invest at a better initial loss. With the script running continuously I could maybe figure out a way to incorporate modeling cross-currency stock trading instead of buying coin outright; there's at least less security risk with dealing with a middle-man. I have a lot to learn about stock trading before I would feel comfortable investing anything but pocket lint in the variability and chance I see in Bitcoin currently, though. Maybe additional research into the market as a whole will reveal the ideal time and stock to enter the game. Again, feel free to use the AppleScript as you wish for your own personal knowledge. It was cool to learn how to automate something and it is still interesting to pick away at for the time being at least. I don't think I will continuously upload unless I see an interest in the content. And this project won't become any more professional over time I'm pretty sure so don't expect me to jump through any real hoops unless there is some kind of real return for the work.